pyproject.toml
[project]
name = "data-alchemy"
version = "0.1.0"
description = "LoRA fine-tuning sample for AMD AI Max+ 395 (ROCm)"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "openai",
    "python-dotenv>=1.2.1",
    "apscheduler",
    "beautifulsoup4",
    "python-docx",
    "pypdf",
    "torch",
    "torchvision",
    "torchaudio",
    "transformers>=4.36.0",
    "tokenizers>=0.15.0",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "accelerate>=0.26.0",
    "sentence-transformers>=2.3.0",
    "faiss-cpu",
    "langchain-community",
    "langchain-core",
    "tiktoken",
    "rank-bm25",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "cryptography>=43.0.0",
    "nest_asyncio>=1.6.0",
    "boto3>=1.34.0",
    "websockets>=12.0",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart",
    "kubernetes>=28.1.0",
    "kopf>=1.36.0",
    "jieba>=0.42.1",
    "polars>=1.0.0",
    "scipy>=1.12.0",
    "s3fs>=2024.2.0",
    "pyspark==3.5.0",
    "triton ; sys_platform == 'linux'",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.6",
]

[project.scripts]
data-alchemy = "src.run_agents:main"
train-lora = "src.train:train"
clean-data = "src.etl.main:main"
schedule-sync = "src.run_agents:main"

[tool.uv]
package = true

[tool.uv.sources]
# Linux ROCm 7.1 support - use AMD official wheels for AI Max+395
# Source: https://repo.radeon.com/rocm/manylinux/rocm-rel-7.1/
# These wheels support AMD AI Max+395 with ROCm 7.1
# Using actual filenames found in AMD repository
# Found via: ./scripts/find_pytorch_wheels.sh
torch = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.1/torch-2.9.1+rocm7.1.0.lw.git351ff442-cp312-cp312-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
  { index = "pytorch", marker = "sys_platform != 'linux'" }
]
torchvision = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.1/torchvision-0.24.0+rocm7.1.0.gitb919bd0c-cp312-cp312-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
  { index = "pytorch", marker = "sys_platform != 'linux'" }
]
torchaudio = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.1/torchaudio-2.9.0+rocm7.1.0.gite3c6ee2b-cp312-cp312-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
  { index = "pytorch", marker = "sys_platform != 'linux'" }
]
# triton is a dependency of torch, add it explicitly
# Note: triton may need to be added to dependencies list if torch requires it
triton = [
  { url = "https://repo.radeon.com/rocm/manylinux/rocm-rel-7.1/triton-3.5.1+rocm7.1.0.gita272dfa8-cp312-cp312-linux_x86_64.whl", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src", "agents", "utils", "etl", "rag", "synthesis", "inference"]


